backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
numpy==2.4.3
opencv-python==4.13.0.92
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
